requirements.txt
pandas==2.1.0
streamlit==1.27.0
numpy==1.24.3
openai==1.3.0
langchain==0.0.350
python-dotenv==1.0.0
pydantic==2.5.2 